Ref: componentsのディレクトリ構造を整理しファイルを仕分け #89
