+++ app/javascript/components/pages/OtherProfile.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import SharedPostIndex from "./SharedPostIndex.vue";
+import SharedPostIndex from "../parts/SharedPostIndex.vue";
 export default {
   components: {
     SharedPostIndex,
