+++ app/javascript/components/pages/PostingsDetail.vue
@@ -3,31 +3,61 @@
     <p class="text-h4 pt-5 title font-weight-bold">{{ post.title }}</p>
     <v-divider></v-divider>
     <div class="text-h6 pt-2 font-weight-bold">
-      <p class="float-left grade_range_scene_type ml-1 mr-5">{{ post.grade_range_ja }}</p>
-      <p class="float-left grade_range_scene_type">{{ post.scene_type_ja }}</p>
-
-      <p class="text-right">
-        投稿者:
-        <router-link
-          :to="{ path: `/${post.user_name}` }"
-          style="text-decoration: none"
-        >
-          {{ post.user_name }}
-        </router-link>
-      </p>
+      <span class="box3 ml-1 mr-5">{{ post.grade_range_ja }}</span>
+      <span class="box3">{{ post.scene_type_ja }}</span>
     </div>
-
-    <div class="description">
-      <p class="text">{{ post.description }}</p>
+    <div class="box8">
+      <p class="content">{{ post.description }}</p>
     </div>
-    <div class="content">
+    <div class="box27">
       <span class="box-title">講話本文</span>
-      <p class="text">{{ newLine(post.content) }}</p>
+      <p class="content">{{ newLine(post.content) }}</p>
     </div>
-    <!-- 参考になったボタン -->
-    <LikeButton @setCurrentUser="setCurrentUser" :postId="this.post.id" />
+    <LikeButton
+      @setCurrentUser="setCurrentUser"
+      class="pb-3"
+      :postId="this.post.id"
+    />
+    <div class="pt-4">
+      <v-btn depressed color="success" class="mr-2">
+        <router-link
+          style="text-decoration: none; color: inherit"
+          :to="{ path: `/postings/edit/${post.id}` }"
+          class="font-weight-bold"
+          >編集</router-link
+        >
+      </v-btn>
+      <v-btn
+        depressed
+        color="error"
+        class="font-weight-bold"
+        @click="confirm_dialog = true"
+      >
+        削除
+      </v-btn>
+      <v-dialog v-model="confirm_dialog" max-width="400">
+        <v-card>
+          <v-card-title>
+            <div>確認ダイアログ</div>
+          </v-card-title>
+          <v-card-text>
+            <p>本当に削除しますか？</p>
+          </v-card-text>
 
-    <v-divider class="mt-10"></v-divider>
+          <v-card-actions>
+            <v-spacer></v-spacer>
+            <v-btn @click="confirm_dialog = false">いいえ、削除しません</v-btn>
+            <v-btn @click="deletePost(post.id)" color="error"
+              >はい、削除します</v-btn
+            >
+          </v-card-actions>
+        </v-card>
+      </v-dialog>
+
+      <!-- <v-btn depressed color="error" v-on:click="deletePost(post.id)">
+        削除
+      </v-btn> -->
+    </div>
 
     <!-- コメント欄↓ -->
     <p class="text-h6 pt-5 pb-3 title font-weight-bold">
@@ -39,9 +69,8 @@
     <p v-if="!comments.length" class="text-h8 font-weight-bold">
       現在コメントはありません。
     </p>
-
     <!-- コメント一覧表示 -->
-    <v-cantainer>
+    <v-container>
       <v-row class="pb-10">
         <v-col
           v-for="comment in this.comments"
@@ -57,7 +86,7 @@
             <v-icon class="pb-1">mdi-clock-outline</v-icon
             >{{ formatDate(comment.updated_at) }}
           </p>
-          <div class="comment">
+          <div class="body">
             <p class="mt-n2">
               {{ newLine(comment.body) }}
             </p>
@@ -65,7 +94,7 @@
 
           <!-- コメント編集フォーム -->
           <div
-            class="text-right pr-8 pt-1"
+            class="text-right pr-9 pt-1"
             v-if="current_user_id === comment.user_id"
           >
             <comment-edit-form
@@ -80,14 +109,12 @@
               color="error"
               class="font-weight-bold"
               @click.stop="confirm_dialog(comment)"
-              id="delete-button"
             >
               削除
             </v-btn>
           </div>
           <v-divider class="mt-3"></v-divider> </v-col
       ></v-row>
-
       <!-- コメント削除用ダイアログ -->
       <v-dialog
         v-model="dialog"
@@ -113,24 +140,20 @@
         </v-card>
       </v-dialog>
       <!-- コメント投稿フォーム -->
-      <CommentForm
-        v-if="loggedIn"
-        @createComment="fetchComments"
-        :postId="this.post.id"
-      />
-    </v-cantainer>
+      <CommentForm @createComment="fetchComments" :postId="this.post.id" />
+    </v-container>
   </v-container>
 </template>
 
 <script>
 import { mapState } from "vuex";
 import { mapActions } from "vuex";
-import LikeButton from "./LikeButton.vue";
-import CommentForm from "./CommentForm.vue";
-import CommentEditForm from "./CommentEditForm.vue";
+import LikeButton from "../parts/LikeButton.vue";
+import CommentForm from "../parts/CommentForm.vue";
+import CommentEditForm from "../parts/CommentEditForm.vue";
 import dayjs from "dayjs";
 export default {
-  name: "PostDetail",
+  name: "PostingsDetail",
   components: {
     LikeButton,
     CommentForm,
@@ -141,34 +164,42 @@ export default {
       post: [],
       comments: [],
       currentComment: {},
-      current_user_id: null,
       dialog: false,
+      current_user_id: null,
+      confirm_dialog: false,
     };
   },
   computed: {
     ...mapState("auth", {
       headers: (state) => state.headers,
-      loggedIn: (state) => state.loggedIn,
     }),
   },
   mounted: function () {
-    this.fetchPostDetail();
+    this.fetchPostingsDetail();
     this.fetchComments();
   },
   methods: {
     ...mapActions("message", ["showMessage"]),
-    fetchPostDetail: function () {
+    fetchPostingsDetail: function () {
       var id = this.$route.params.id;
-      this.$axios.get("/posts/" + id).then(
-        (response) => {
-          this.post = response.data;
-          console.log("取得完了");
-          // this.user = response.data;
-        },
-        (error) => {
-          console.log(error);
-        }
-      );
+      this.$axios
+        .get("/postings/" + id, {
+          headers: {
+            uid: this.headers["uid"],
+            "access-token": this.headers["access-token"],
+            client: this.headers["client"],
+          },
+        })
+        .then(
+          (response) => {
+            this.post = response.data;
+            console.log("取得完了");
+            // this.user = response.data;
+          },
+          (error) => {
+            console.log(error);
+          }
+        );
     },
     fetchComments: function () {
       var id = this.$route.params.id;
@@ -186,9 +217,34 @@ export default {
       console.log("setCurrentUserUser発動");
       console.log(this.current_user_id);
     },
-    confirm_dialog(comment) {
+    confirm_comment_dialog(comment) {
       this.dialog = true;
       this.currentComment = comment;
+    },
+    deletePost: function () {
+      var id = this.$route.params.id;
+      this.$axios
+        .delete("/postings/" + id, {
+          headers: {
+            uid: this.headers["uid"],
+            "access-token": this.headers["access-token"],
+            client: this.headers["client"],
+          },
+        })
+        .then(
+          (response) => {
+            this.confirm_dialog = false;
+            this.showMessage({
+              message: "講話を削除しました。",
+              type: "warning",
+              status: true,
+            }),
+              this.$router.push({ name: "Postings" });
+          },
+          (error) => {
+            console.log(error);
+          }
+        );
     },
     deleteComment(id) {
       this.$axios
@@ -223,7 +279,7 @@ export default {
 </script>
 
 <style scoped>
-.description {
+.box8 {
   padding: 0.5em 1em;
   margin: 2em 0;
   color: #232323;
@@ -231,25 +287,23 @@ export default {
   border-left: solid 10px #7d420aa4;
   border-radius: 5px;
 }
-.grade_range_scene_type {
-  padding: 0em 0.5em;
+.box3 {
+  padding: 0.5em 0.5em;
+  /* margin: 2em 0; */
   color: #2c2c2f;
-  background: #efefef;
+  background: #efefef; /*背景色*/
   border-radius: 30px;
 }
-/* .heading {
-  border-bottom: 4px double #000;
-} */
 
-.content {
+.box27 {
   position: relative;
-  margin: 3em 0;
+  margin: 3em 0 2em 0;
   padding: 0.5em 1em;
   border: solid 3px #7d420aa4;
   background: #eafff0;
   border-radius: 5px;
 }
-.content .box-title {
+.box27 .box-title {
   position: absolute;
   display: inline-block;
   top: -27px;
@@ -263,19 +317,26 @@ export default {
   font-weight: bold;
   border-radius: 5px 5px 0 0;
 }
-
-.text {
+.box27 p {
+  margin: 0;
+  padding: 0;
+}
+.content {
   white-space: pre-line;
   font-size: 19px;
 }
-.comment {
+.body {
+  white-space: pre-line;
+  font-size: 19px;
+}
+.body {
   padding: 0.5em 1em;
   margin: 0 2em 0 0;
   color: #2c2c2f;
   background: #e4ecf5; /*背景色*/
   border-radius: 5px;
 }
-.comment p {
+.body p {
   white-space: pre-line;
 }
 </style>
