+++ app/javascript/components/pages/PrivacyPolicy.vue
