+++ app/javascript/components/pages/Terms.vue
