+++ app/javascript/components/parts/SharedPostIndex.vue
